Refactored Component As Feature Based Folders
