+++ src/components/searchbox/SearchBox.tsx
@@ -1,0 +1,20 @@
+import {BoxArea108} from './index.ts';
+
+export default  function BoxArea97() {
+    const onSearch = (search: string) => {
+        console.log(search);
+        // implementing the search logic is not required for this task
+    };
+
+    return (
+        <div id={'box-area'} className="relative w-full max-w-5xl mx-auto rounded-xl overflow-hidden mt-8">
+            <img src="/task1/hero-bg.png" className="w-full h-96 object-cover" />
+            <div className="absolute inset-0 bg-black/50 flex flex-col items-center justify-center text-center px-4">
+                <h1 className="text-3xl md:text-5xl font-bold text-white">
+                    Search for words, phrases and meanings
+                </h1>
+                <BoxArea108 initialValue="" onSearch={onSearch} />
+            </div>
+        </div>
+    );
+};
