+++ src/components/searchbox/index.ts
@@ -1,0 +1,2 @@
+export { default as BoxArea97 } from "./SearchBox";
+export { default as BoxArea108 } from "./SearchInput";
