+++ src/components/taglist/index.ts
@@ -1,0 +1,1 @@
+export {default as TagList} from './TagList.tsx';
