+++ src/components/taglist/types.ts
@@ -1,0 +1,4 @@
+export interface TagListProps {
+    title: string;
+    tags: string[];
+}
